Feat: 검색 디테일 추가
- 최근일수록 상단에 표시되도록 변경
- 최대 10개까지만 표시
