+++ components/ui/SearchHistoryItem.tsx
@@ -1,18 +1,18 @@
-import { recentSearchKeyword, useRecent } from "@/state/recentKeywordState";
+import { useRecent } from "@/hooks/useRecent";
 import { useRouter } from "next/router";
 import React from "react";
-import { useRecoilState } from "recoil";
 
 type SearchedItem = {
   item?: string;
 };
 
-export default function LatestSearchItem({ item }: SearchedItem) {
+export default function LatestSearchItem({ item = "" }: SearchedItem) {
   const router = useRouter();
 
   const [recentSearchKeywords, setRecentSearchKeywords] = useRecent();
 
-  const onClickHandler = () =>
+  const onClickHandler = () => {
+    putKeyword(item);
     router.push({
       pathname: "/search_result",
       query: {
@@ -20,11 +20,32 @@ export default function LatestSearchItem({ item }: SearchedItem) {
         bigCategory: `${"전체"}`,
       },
     });
+  };
 
   const deleteHandler = () => {
     const res = recentSearchKeywords.filter((k) => k !== item);
     setRecentSearchKeywords([...res]);
   };
+
+  function putKeyword(keyword: string) {
+    //이미 있는 키워드 재검색 시 가장 앞으로 갱신
+    if (recentSearchKeywords.includes(keyword)) {
+      setRecentSearchKeywords([
+        keyword,
+        ...recentSearchKeywords.filter((k) => k !== keyword),
+      ]);
+    } else {
+      // 없을 경우 앞에 추가
+      setRecentSearchKeywords([keyword, ...recentSearchKeywords]);
+    }
+
+    //추가했는데 개수가 10개 초과인 경우 제일 오래된 값 제거
+    if (recentSearchKeywords.length >= 10) {
+      let keywords: string[] = [keyword, ...recentSearchKeywords];
+      keywords.pop();
+      setRecentSearchKeywords([...keywords]);
+    }
+  }
 
   return (
     <div className="keywords">
